Java Assignment 2 - Website Login with Comment
This is a website made with Java. It allows you to login with an Admin
account. The user has an option to create a new user which will be
stored into a database. Once the users registers and logins, they will
go directed to a page where they can go and post comments on the
website.